requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-multipart==0.0.17
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
